+++ controllers/__init__.py
@@ -1,0 +1,1 @@
+# Pacote de controllers da aplicação.
